+++ src/app/components/dialog-text/dialog-text.component.html
@@ -8,17 +8,17 @@
 <br>
 <br>
 <p> {{this.multilinguism.translate('speechToText')}} </p>
-<div *ngIf="voiceActiveSectionDisabled; else voicesection;">
-  <button class="backgroundWhite" mat-raised-button type="button" (click)="startVoiceRecognition()"><img height="60px" width="50px" src="assets/icons/microphone.svg" alt="microphone"/></button>
+<div *ngIf="this.voiceRecognition.voiceActiveSectionDisabled; else voicesection;">
+  <button class="backgroundWhite" mat-raised-button type="button" (click)="this.voiceRecognition.startVoiceRecognition()"><img height="60px" width="50px" src="assets/icons/microphone.svg" alt="microphone"/></button>
 </div>
 <ng-template #voicesection>
-  <ng-container *ngIf="!voiceActiveSectionError; else failure">
-    <ng-container *ngIf="voiceText; else start">
-      <span>{{voiceText}}</span>
+  <ng-container *ngIf="!this.voiceRecognition.voiceActiveSectionError; else failure">
+    <ng-container *ngIf="this.voiceRecognition.voiceText; else start">
+      <span>{{this.voiceRecognition.voiceText}}</span>
     </ng-container>
     <ng-template #start>
       <ng-container
-        *ngIf="voiceActiveSectionListening; else beginning">
+        *ngIf="this.voiceRecognition.voiceActiveSectionListening; else beginning">
         <span>{{this.multilinguism.translate('Listening')}}...</span>
       </ng-container>
       <ng-template #beginning>
@@ -30,8 +30,8 @@
     <span>{{this.multilinguism.translate('Didnt catch that')}}</span>
   </ng-template>
   <div>
-    <button mat-raised-button color="warn" (click)="closeVoiceRecognition()">{{this.multilinguism.translate('Close')}}</button>
-    <button mat-raised-button color="primary" (click)="startVoiceRecognition()">{{this.multilinguism.translate('Restart')}}</button>
-    <button mat-raised-button color="accent" *ngIf="voiceText" (click)="submitFromSpeech()" mat-dialog-close>{{this.multilinguism.translate('submit')}}</button>
+    <button mat-raised-button color="warn" (click)="this.voiceRecognition.closeVoiceRecognition()">{{this.multilinguism.translate('Close')}}</button>
+    <button mat-raised-button color="primary" (click)="this.voiceRecognition.startVoiceRecognition()">{{this.multilinguism.translate('Restart')}}</button>
+    <button mat-raised-button color="accent" *ngIf="this.voiceRecognition.voiceText" (click)="submitFromSpeech()" mat-dialog-close>{{this.multilinguism.translate('submit')}}</button>
   </div>
 </ng-template>
