+++ src/app/components/dialog-text/dialog-text.component.ts
@@ -5,6 +5,7 @@ import {MultilinguismService} from "../../services/multilinguism.service";
 import {ConfigurationService} from "../../services/configuration.service";
 import {UsertoolbarService} from "../../services/usertoolbar.service";
 import {GeticonService} from "../../services/geticon.service";
+import {VoiceRecognitionService} from "../../services/voice-recognition.service";
 
 declare const annyang: any;
 
@@ -16,18 +17,12 @@ declare const annyang: any;
 export class DialogTextComponent implements OnInit {
   private name: any;
 
-  voiceActiveSectionDisabled: boolean = true;
-  voiceActiveSectionError: boolean = false;
-  voiceActiveSectionSuccess: boolean = false;
-  voiceActiveSectionListening: boolean = false;
-  voiceText: any;
-
   constructor(private historicService: HistoricService,
               public multilinguism: MultilinguismService,
               private configurationService: ConfigurationService,
               public getIconService: GeticonService,
               public userToolBarService: UsertoolbarService,
-              private ngZone: NgZone) { }
+              public voiceRecognition: VoiceRecognitionService) { }
 
   ngOnInit(): void {
     this.changeLanguage();
@@ -49,7 +44,7 @@ export class DialogTextComponent implements OnInit {
 
   submitFromSpeech() {
     const vignette: Vignette = {
-      Label: this.voiceText,
+      Label: this.voiceRecognition.voiceText,
       ImagePath: '',
       Color: '',
       BorderColor: '',
@@ -57,73 +52,6 @@ export class DialogTextComponent implements OnInit {
     this.historicService.push(vignette);
   }
 
-  initializeVoiceRecognitionCallback(): void {
-    annyang.addCallback('error', (err) => {
-      if(err.error === 'network'){
-        this.voiceText = "Internet is require";
-        annyang.abort();
-        this.ngZone.run(() => this.voiceActiveSectionSuccess = true);
-      } else if (this.voiceText === undefined) {
-        this.ngZone.run(() => this.voiceActiveSectionError = true);
-        annyang.abort();
-      }
-    });
-
-    annyang.addCallback('soundstart', (res) => {
-      this.ngZone.run(() => this.voiceActiveSectionListening = true);
-    });
-
-    annyang.addCallback('end', () => {
-      if (this.voiceText === undefined) {
-        this.ngZone.run(() => this.voiceActiveSectionError = true);
-        annyang.abort();
-      }
-    });
-
-    annyang.addCallback('result', (userSaid) => {
-      this.ngZone.run(() => this.voiceActiveSectionError = false);
-
-      let queryText: any = userSaid[0];
-
-      annyang.abort();
-
-      this.voiceText = queryText;
-
-      this.ngZone.run(() => this.voiceActiveSectionListening = false);
-      this.ngZone.run(() => this.voiceActiveSectionSuccess = true);
-    });
-  }
-
-  startVoiceRecognition(): void {
-    this.voiceActiveSectionDisabled = false;
-    this.voiceActiveSectionError = false;
-    this.voiceActiveSectionSuccess = false;
-    this.voiceText = undefined;
-
-    if (annyang) {
-      let commands = {
-        'demo-annyang': () => { }
-      };
-
-      annyang.addCommands(commands);
-
-      this.initializeVoiceRecognitionCallback();
-
-      annyang.start({ autoRestart: false });
-    }
-  }
-
-  closeVoiceRecognition(): void {
-    this.voiceActiveSectionDisabled = true;
-    this.voiceActiveSectionError = false;
-    this.voiceActiveSectionSuccess = false;
-    this.voiceActiveSectionListening = false;
-    this.voiceText = undefined;
-
-    if(annyang){
-      annyang.abort();
-    }
-  }
   changeLanguage(){
     if(this.configurationService.LANGUAGE_VALUE === 'FR'){
       annyang.setLanguage('fr-FR');
@@ -132,6 +60,7 @@ export class DialogTextComponent implements OnInit {
       annyang.setLanguage('en');
     }
   }
+
   getIcon(s: string) {
     return this.getIconService.getIconUrl(s);
   }
