+++ src/app/services/voice-recognition.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { VoiceRecognitionService } from './voice-recognition.service';
+
+describe('VoiceRecognitionService', () => {
+  let service: VoiceRecognitionService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(VoiceRecognitionService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
